move the listener inside the wizard ui
